# internal/handler/openapi_embed.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RideMatch API",
    "description": "Backend API for RideMatch — a multi-modal (car, okada, keke, bus) ride/logistics matching platform for Nigeria. Authentication is phone number + OTP based; there is no password field. All responses use a single envelope shape: `{ success, message, data, error }`.",
    "version": "1.0.0",
    "contact": {
      "name": "Joshua Olumoye"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token returned from /auth/otp/verify or /auth/token/refresh. Send as `Authorization: Bearer <token>`."
      }
    },
    "schemas": {
      "APIResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "error": {
            "type": "string"
          }
        }
      },
...
      "RegisterDriverRequest": {
        "type": "object",
        "required": [
          "vehicle_type",
          "plate_number",
          "id_document_url"
        ],
        "properties": {
          "vehicle_type": {
            "type": "string",
            "enum": [
              "car",
              "okada",
              "keke",
              "bus"
            ],
            "example": "okada"
          },
          "plate_number": {
            "type": "string",
            "example": "KJA-123-XY"
          },
          "vehicle_photo_url": {
            "type": "string",
            "format": "uri"
          },
          "id_document_url": {
            "type": "string",
            "format": "uri",
            "description": "Government ID or driver's license photo, hosted at a URL reachable by the backend. Never returned by any read endpoint."
          }
        }
      },
      "DriverProfileResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "vehicle_type": {
            "type": "string",
            "enum": [
              "car",
              "okada",
              "keke",
              "bus"
            ]
          },
          "plate_number": {
            "type": "string"
          },
          "vehicle_photo_url": {
            "type": "string"
          },
          "verification_status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "verification_note": {
            "type": "string"
          },
          "is_online": {
            "type": "boolean"
          },
          "has_active_subscription": {
            "type": "boolean"
          },
          "subscription_active_until": {
            "type": "string",
            "format": "date-time"
          },
          "rating_average": {
            "type": "number",
            "format": "float"
          }
        }
      },
      "GoOnlineRequest": {
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "properties": {
          "latitude": {
            "type": "number",
            "format": "float",
            "example": 6.5244
          },
          "longitude": {
            "type": "number",
            "format": "float",
            "example": 3.3792
          }
        }
      },
      "LocationPingRequest": {
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "properties": {
          "latitude": {
            "type": "number",
            "format": "float",
            "example": 6.5244
          },
          "longitude": {
            "type": "number",
            "format": "float",
            "example": 3.3792
          }
        }
      },
      "NearbyDriverResponse": {
        "type": "object",
        "properties": {
          "driver_id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "photo_url": {
            "type": "string"
          },
          "vehicle_type": {
            "type": "string",
            "enum": [
              "car",
              "okada",
              "keke",
              "bus"
            ]
          },
          "rating_average": {
            "type": "number",
            "format": "float"
          },
          "distance_km": {
            "type": "number",
            "format": "float"
          },
          "latitude": {
            "type": "number",
            "format": "float",
            "description": "Rounded, not exact — see NEARBY_COORDINATE_PRECISION"
          },
          "longitude": {
            "type": "number",
            "format": "float"
          }
        }
      },
...
      "InitiateSubscriptionRequest": {
        "type": "object",
        "properties": {
          "days": {
            "type": "integer",
            "minimum": 1,
            "maximum": 30,
            "default": 1,
            "description": "Days of platform access to pay for"
          },
          "redirect_url": {
            "type": "string",
            "format": "uri",
            "description": "Where Flutterwave sends the driver after payment (defaults to FLW_REDIRECT_URL)"
          }
        }
      },
      "InitiateSubscriptionResponse": {
        "type": "object",
        "properties": {
          "payment_link": {
            "type": "string",
            "format": "uri",
            "description": "Open this in a browser/WebView to complete payment"
          },
          "tx_ref": {
            "type": "string"
          },
          "amount_kobo": {
            "type": "integer",
            "format": "int64"
          },
          "days": {
            "type": "integer"
          }
        }
      },
      "FlutterwaveWebhookPayload": {
        "type": "object",
        "description": "Sent by Flutterwave, not constructed by API clients. Shown here for documentation completeness.",
        "properties": {
          "event": {
            "type": "string",
            "example": "charge.completed"
          },
          "data": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "format": "int64"
              },
              "tx_ref": {
                "type": "string"
              },
              "status": {
                "type": "string",
                "example": "successful"
              },
              "amount": {
                "type": "number",
                "format": "float"
              },
              "currency": {
                "type": "string",
                "example": "NGN"
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/driver/register": {
      "post": {
        "tags": [
          "Driver"
        ],
        "summary": "Register as a driver",
        "description": "Creates a driver profile for the authenticated account and promotes it to the \"driver\" role. Vehicle photo and ID document must already be hosted somewhere reachable by URL. Starts in \"pending\" verification status.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDriverRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Driver profile created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DriverProfileResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing/invalid fields"
          },
          "409": {
            "description": "Already registered, or plate number taken"
          }
        }
      }
    },
    "/driver/profile": {
      "get": {
        "tags": [
          "Driver"
        ],
        "summary": "Get the authenticated driver's profile",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DriverProfileResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Not registered as a driver"
          }
        }
      }
    },
    "/driver/online": {
      "post": {
        "tags": [
          "Driver"
        ],
        "summary": "Go online",
        "description": "Marks the driver online and publishes their live position so they appear in nearby-driver searches. Requires an approved driver profile and an active subscription.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoOnlineRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Now online",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/DriverProfileResponse"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Not verified, or subscription inactive"
          },
          "404": {
            "description": "Not registered as a driver"
          }
        }
      }
    },
    "/driver/offline": {
      "post": {
        "tags": [
          "Driver"
        ],
        "summary": "Go offline",
        "description": "Marks the driver offline and removes them from nearby-driver searches.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Now offline"
          },
          "404": {
            "description": "Not registered as a driver"
          }
        }
      }
    },
    "/driver/location": {
      "post": {
        "tags": [
          "Driver"
        ],
        "summary": "Send a location update",
        "description": "Updates the driver's live position. Call periodically (every 5-10 seconds) while online — a driver who stops pinging is treated as stale and dropped from nearby-driver results after the configured timeout (default 45s).",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationPingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location updated"
          },
          "400": {
            "description": "Driver is not currently online"
          }
        }
      }
    },
    "/drivers/nearby": {
      "get": {
        "tags": [
          "Driver"
        ],
        "summary": "Find nearby online drivers",
        "description": "Passenger-facing: returns online drivers of the given vehicle type near a point, nearest first. Coordinates in the response are rounded, not exact, until a match is confirmed.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "vehicle_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "car",
                "okada",
                "keke",
                "bus"
              ]
            }
          },
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "format": "float"
            }
          },
          {
            "name": "radius_km",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "format": "float",
              "default": 5
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Nearby drivers",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/NearbyDriverResponse"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing/invalid query parameters"
          }
        }
      }
    },
...
    "/driver/subscription/checkout": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Start a driver subscription payment",
        "description": "Creates a Flutterwave hosted-checkout link for N days of platform access (default 1, max 30). Nothing is granted yet — access activates only once the payment webhook confirms the charge; poll GET /driver/profile afterward to see `has_active_subscription` flip to true.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateSubscriptionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Checkout link created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/APIResponse"
                    }
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/InitiateSubscriptionResponse"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not registered as a driver yet"
          }
        }
      }
    },
    "/webhooks/flutterwave": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Flutterwave payment webhook",
        "description": "Called by Flutterwave when a charge completes. Verified two ways before anything is credited: the `verif-hash` header must match FLW_WEBHOOK_SECRET_HASH, and the transaction is independently re-fetched from Flutterwave's API before the driver's subscription is activated. Configure this URL as your webhook endpoint in the Flutterwave dashboard. Not for use by API clients directly.",
        "parameters": [
          {
            "name": "verif-hash",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Shared secret configured in the Flutterwave dashboard"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlutterwaveWebhookPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed (or safely ignored, e.g. an unhandled event type)"
          },
          "401": {
            "description": "Invalid or missing verif-hash"
          }
        }
      }
    }
  }
}